+++ native/packages/react-native-bpk-component-button/src/BpkButton.js
@@ -23,7 +23,9 @@
  import React from 'react';
  import PropTypes from 'prop-types';
  import LinearGradient from 'react-native-linear-gradient';
+ import xor from 'lodash/xor';
 
+ import { withTheme } from 'react-native-bpk-theming';
  import BpkText from 'react-native-bpk-component-text';
 
  import styles from './BpkButton-styles';
@@ -64,8 +66,24 @@
    return styleForElement;
  };
 
- const getGradientColors = ({ type, disabled, selected }) => {
-   let gradientColors = styles.gradientColors[type];
+ const getThemingForElement = (elementType, theme, { type, selected, disabled }) => {
+   const themeForElement = {};
+   if (theme && !selected && !disabled && styles.themeMappings[elementType]) {
+     Object.keys(styles.themeMappings[elementType]).forEach((key) => {
+       const values = styles.themeMappings[elementType][key];
+       if (values[type]) {
+         themeForElement[key] = theme[values[type]];
+       }
+     });
+   }
+   return themeForElement;
+ };
+
+ const getGradientColors = (theme, { type, disabled, selected }) => {
+   let gradientColors = theme ?
+   [theme.buttonPrimaryGradientStartColor, theme.buttonPrimaryGradientStartColor] :
+   styles.gradientColors[type];
+
    if (selected) {
      gradientColors = styles.gradientColors.selected;
    }
@@ -97,9 +115,29 @@
      style,
      ...rest
    } = props;
+   let theme = props.theme;
 
+   // Validate the button type.
    if (!BUTTON_TYPES.includes(type)) {
      throw new Error(`"${type}" is not a valid button type. Valid types are ${BUTTON_TYPES.join(', ')}`);
+   }
+
+   // Validate that correct theming attributes have been provided.
+   if (theme) {
+     // Must provide disabled and selected states regardless of theme.
+     const attributesToSupply = [
+       // TODO this should support multiple types in future with some logic.
+       styles.themeMappings.text.color.primary,
+       styles.themeMappings.gradient.primary.startColor,
+       styles.themeMappings.gradient.primary.endColor,
+     ];
+
+     // If all attributes were not supplied, turn off theming.
+     // TODO allow all types to be themed.
+     const allAttributesSupplied = xor(attributesToSupply, Object.keys(theme)).length === 0;
+     if (!allAttributesSupplied || type !== 'primary') {
+       theme = null;
+     }
    }
 
    const accessibilityTraits = ['button'];
@@ -110,7 +148,10 @@
    // Note that TouchableHighlight isn't on Android, so TouchableFeedback
    // will need to be used to support it.
    return (
-     <LinearGradient style={[getStyleForElement('container', props), style]} colors={getGradientColors(props)}>
+     <LinearGradient
+       colors={getGradientColors(theme, props)}
+       style={[getStyleForElement('container', props), style]}
+     >
        <TouchableHighlight
          style={getStyleForElement('button', props)}
          disabled={disabled}
@@ -127,7 +168,7 @@
              <BpkText
                textStyle={large ? 'lg' : 'sm'}
                emphasize
-               style={getStyleForElement('text', props)}
+               style={[getStyleForElement('text', props), getThemingForElement('text', theme, props)]}
              >
                {title}
              </BpkText>
@@ -139,7 +180,7 @@
    );
  };
 
- BpkButton.propTypes = {
+ const propTypes = {
    title: PropTypes.string.isRequired,
    icon: iconPropType,
    iconOnly: PropTypes.bool,
@@ -150,7 +191,14 @@
    selected: PropTypes.bool,
    accessibilityLabel: PropTypes.string,
    style: View.propTypes.style,
+   theme: PropTypes.shape({
+     gradientStartColor: PropTypes.string.isRequired,
+     gradientEndColor: PropTypes.string.isRequired,
+     textColor: PropTypes.string.isRequired,
+   }),
  };
+
+ BpkButton.propTypes = propTypes;
 
  BpkButton.defaultProps = {
    icon: null,
@@ -161,7 +209,11 @@
    selected: false,
    accessibilityLabel: null,
    style: null,
+   theme: null,
  };
 
- export default BpkButton;
- export { BUTTON_TYPES };
+ export default withTheme(BpkButton);
+ export {
+   propTypes,
+   BUTTON_TYPES,
+  };