+++ changelog.md
@@ -7,6 +7,8 @@
   - iOS & Android border size tokens
   - Android elevation tokens
   - iOS touchable highlight underlay tokens
+- react-native-bpk-component-button
+  - Added theming support
 
 **Fixed:**
 - bpk-tokens: